Creando funcionalidad del mapa. close #21, close #23, close #24, close #25

+++ app/src/main/java/com/alvar/practica7pmdm/DataBaseManager/Esquema.java
@@ -14,15 +14,14 @@ public class Esquema {
         public static final String COLUMN_NAME_NOMBRE = "nombre";
         public static final String COLUMN_NAME_LATITUD = "latitud";
         public static final String COLUMN_NAME_LONGITUD = "longitud";
-
         public static final String COLUMN_NAME_COMENTARIOS = "comentarios";
         public static final String COLUMN_NAME_VALORACION = "valoracion";
         public static final String COLUMN_NAME_CATEGORIA = "categoria";
 
         public static final String COLUMN_TYPE_ID = "INTEGER";
         public static final String COLUMN_TYPE_NOMBRE = "TEXT";
-        public static final String COLUMN_TYPE_LATITUD = "FLOAT";
-        public static final String COLUMN_TYPE_LONGITUD = "FLOAT";
+        public static final String COLUMN_TYPE_LATITUD = "DOUBLE";
+        public static final String COLUMN_TYPE_LONGITUD = "DOUBLE";
         public static final String COLUMN_TYPE_COMENTARIOS = "TEXT";
         public static final String COLUMN_TYPE_VALORACION = "FLOAT";
         public static final String COLUMN_TYPE_CATEGORIA = "INT";
